refacotor: moving test porj in test-utils
