+++ src/generate/utils_spec.ts
@@ -2,7 +2,7 @@ import { join } from 'path';
 
 import { SchematicTestRunner } from '@angular-devkit/schematics/testing';
 
-import { createEmptyNsOnlyProject, createEmptySharedProject, toComponentClassName } from '../utils';
+import { createEmptyNsOnlyProject, createEmptySharedProject } from '../test-utils';
 import { Schema as ComponentOptions } from './component/schema';
 import { Schema as ModuleOptions } from './module/schema';
 import { getPlatformUse } from './utils';
